Authentication added

+++ src/main/java/com/sedbproj/Railways/Repositories/StationRepository.java
@@ -3,5 +3,5 @@ package com.sedbproj.Railways.Repositories;
 import com.sedbproj.Railways.Entities.StationEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface StationRepository extends CrudRepository<StationEntity, Long> {
+public interface StationRepository extends CrudRepository<StationEntity, Integer> {
 }
